Catch diagnostics error and split out tests
